in-target: accept --pass-stdout option and pass it on to log-output
Allows more flexible use of in-target, especially for single commands
that need a correct target environment.

The change has been tested, but is not yet used as I decided on a
different solution for the use-case I wrote the patch for. Committing
anyway as it seems a useful feature.

r53511